security-service/requirements.txt
# security-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.4.2
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-exporter-otlp-proto-grpc==1.36.0